Recombine Context2D in to Canvas widget
