+++ src/core/toga/__init__.py
@@ -13,7 +13,6 @@ from .widgets.box import Box
 from .widgets.button import Button
 from .widgets.detailedlist import DetailedList
 from .widgets.canvas import Canvas
-from .widgets.context2d import Context2D
 # from .widgets.icon import Icon
 from .widgets.image import *
 from .widgets.imageview import *
