+++ src/gtk/toga_gtk/factory.py
@@ -7,7 +7,6 @@ from .widgets.box import Box
 from .widgets.button import Button
 from .command import Command
 from .widgets.canvas import Canvas
-from .widgets.context2d import Context2D
 from .widgets.icon import Icon
 # from .widgets.image import *
 # from .widgets.imageview import *
